# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Website Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ServerStatus = (typeof ServerStatus)[keyof typeof ServerStatus];

export const ServerStatus = {
  connected: "connected",
  disconnected: "disconnected",
  unknown: "unknown",
} as const;

export interface Server {
  id: number;
  name: string;
  host: string;
  port: number;
  username: string;
  status: ServerStatus;
  nginxInstalled: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface CreateServerBody {
  name: string;
  host: string;
  port: number;
  username: string;
  password: string;
  /** @nullable */
  privateKey?: string | null;
}

export interface UpdateServerBody {
  name?: string;
  host?: string;
  port?: number;
  username?: string;
  password?: string;
  /** @nullable */
  privateKey?: string | null;
}

export interface ConnectionTestResult {
  success: boolean;
  message: string;
  /** @nullable */
  output?: string | null;
}

export interface CommandResult {
  success: boolean;
  output: string;
  /** @nullable */
  error?: string | null;
}

export interface ServerStats {
  cpu: number;
  memoryUsed: number;
  memoryTotal: number;
  diskUsed: number;
  diskTotal: number;
  uptime: string;
}

export type SiteSiteType = (typeof SiteSiteType)[keyof typeof SiteSiteType];

export const SiteSiteType = {
  static: "static",
  nodejs: "nodejs",
  php: "php",
  python: "python",
} as const;

export type SiteStatus = (typeof SiteStatus)[keyof typeof SiteStatus];

export const SiteStatus = {
  deploying: "deploying",
  active: "active",
  failed: "failed",
  stopped: "stopped",
} as const;

export interface Site {
  id: number;
  serverId: number;
  name: string;
  domain: string;
  /** @nullable */
  repoUrl?: string | null;
  /** @nullable */
  repoToken?: string | null;
  deployPath: string;
  /** @nullable */
  buildCommand?: string | null;
  siteType: SiteSiteType;
  status: SiteStatus;
  sslInstalled: boolean;
  autoSync: boolean;
  /** @nullable */
  lastDeployedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateSiteBodySiteType =
  (typeof CreateSiteBodySiteType)[keyof typeof CreateSiteBodySiteType];

export const CreateSiteBodySiteType = {
  static: "static",
  nodejs: "nodejs",
  php: "php",
  python: "python",
} as const;

export interface CreateSiteBody {
  serverId: number;
  name: string;
  domain: string;
  /** @nullable */
  repoUrl?: string | null;
  /** @nullable */
  repoToken?: string | null;
  deployPath: string;
  /** @nullable */
  buildCommand?: string | null;
  siteType: CreateSiteBodySiteType;
  autoSync?: boolean;
}

export type UpdateSiteBodySiteType =
  (typeof UpdateSiteBodySiteType)[keyof typeof UpdateSiteBodySiteType];

export const UpdateSiteBodySiteType = {
  static: "static",
  nodejs: "nodejs",
  php: "php",
  python: "python",
} as const;

export interface UpdateSiteBody {
  name?: string;
  domain?: string;
  /** @nullable */
  repoUrl?: string | null;
  /** @nullable */
  repoToken?: string | null;
  /** @nullable */
  buildCommand?: string | null;
  siteType?: UpdateSiteBodySiteType;
  autoSync?: boolean;
}

export interface CloudflareConfig {
  id: number;
  label: string;
  email: string;
  /** @nullable */
  zoneId?: string | null;
  createdAt: string;
}

export interface CreateCloudflareConfigBody {
  label: string;
  email: string;
  apiToken: string;
  /** @nullable */
  zoneId?: string | null;
}

export interface CloudflareZone {
  id: string;
  name: string;
  status: string;
}

export interface CreateDnsRecordBody {
  zoneId: string;
  domain: string;
  ip: string;
  proxied?: boolean;
}

export type ActivityEntryType =
  (typeof ActivityEntryType)[keyof typeof ActivityEntryType];

export const ActivityEntryType = {
  deploy: "deploy",
  ssl: "ssl",
  nginx_install: "nginx_install",
  dns_setup: "dns_setup",
  connection_test: "connection_test",
  error: "error",
} as const;

export type ActivityEntryStatus =
  (typeof ActivityEntryStatus)[keyof typeof ActivityEntryStatus];

export const ActivityEntryStatus = {
  success: "success",
  failure: "failure",
  running: "running",
} as const;

export interface ActivityEntry {
  id: number;
  /** @nullable */
  siteId?: number | null;
  /** @nullable */
  serverId?: number | null;
  type: ActivityEntryType;
  status: ActivityEntryStatus;
  message: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalServers: number;
  totalSites: number;
  activeSites: number;
  sslEnabled: number;
  recentDeployments: number;
}