# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VPS Website Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List all VPS servers
      responses:
        "200":
          description: List of VPS servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
    post:
      operationId: createServer
      tags: [servers]
      summary: Add a new VPS server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerBody"
      responses:
        "201":
          description: Created server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get a specific VPS server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateServer
      tags: [servers]
      summary: Update a VPS server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerBody"
      responses:
        "200":
          description: Updated server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Remove a VPS server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /servers/{id}/test-connection:
    post:
      operationId: testServerConnection
      tags: [servers]
      summary: Test SSH connection to VPS server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
  /servers/{id}/install-nginx:
    post:
      operationId: installNginx
      tags: [servers]
      summary: Install and configure Nginx on VPS server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Nginx install result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
  /servers/{id}/stats:
    get:
      operationId: getServerStats
      tags: [servers]
      summary: Get server resource stats (CPU, memory, disk)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerStats"
...
  /cloudflare:
    get:
      operationId: listCloudflareConfigs
      tags: [cloudflare]
      summary: List Cloudflare configurations
      responses:
        "200":
          description: Cloudflare configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CloudflareConfig"
    post:
      operationId: createCloudflareConfig
      tags: [cloudflare]
      summary: Save a Cloudflare API configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCloudflareConfigBody"
      responses:
        "201":
          description: Created config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloudflareConfig"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cloudflare/{id}:
    delete:
      operationId: deleteCloudflareConfig
      tags: [cloudflare]
      summary: Delete a Cloudflare config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cloudflare/{id}/zones:
    get:
      operationId: getCloudflareZones
      tags: [cloudflare]
      summary: List zones in a Cloudflare account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CloudflareZone"
  /cloudflare/{id}/create-dns:
    post:
      operationId: createDnsRecord
      tags: [cloudflare]
      summary: Create a DNS A record and enable proxy in Cloudflare
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDnsRecordBody"
      responses:
        "200":
          description: DNS record creation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Server:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        status:
          type: string
          enum: [connected, disconnected, unknown]
        nginxInstalled:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - host
        - port
        - username
        - status
        - nginxInstalled
        - createdAt
        - updatedAt
    CreateServerBody:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        privateKey:
          type: ["string", "null"]
      required:
        - name
        - host
        - port
        - username
        - password
    UpdateServerBody:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        privateKey:
          type: ["string", "null"]
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        output:
          type: ["string", "null"]
      required:
        - success
        - message
    CommandResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        error:
          type: ["string", "null"]
      required:
        - success
        - output
    ServerStats:
      type: object
      properties:
        cpu:
          type: number
        memoryUsed:
          type: number
        memoryTotal:
          type: number
        diskUsed:
          type: number
        diskTotal:
          type: number
        uptime:
          type: string
      required:
        - cpu
        - memoryUsed
        - memoryTotal
        - diskUsed
        - diskTotal
        - uptime
...
    CloudflareConfig:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        email:
          type: string
        zoneId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - label
        - email
        - createdAt
    CreateCloudflareConfigBody:
      type: object
      properties:
        label:
          type: string
        email:
          type: string
        apiToken:
          type: string
        zoneId:
          type: ["string", "null"]
      required:
        - label
        - email
        - apiToken
    CloudflareZone:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
      required:
        - id
        - name
        - status
    CreateDnsRecordBody:
      type: object
      properties:
        zoneId:
          type: string
        domain:
          type: string
        ip:
          type: string
        proxied:
          type: boolean
      required:
        - zoneId
        - domain
        - ip
...