# lib/api-zod/src/generated/types/createCloudflareConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Website Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCloudflareConfigBody {
  label: string;
  email: string;
  apiToken: string;
  /** @nullable */
  zoneId?: string | null;
}